don't return map error when negative now, still need to resolve in draw negative number because it doestn't draw the points

+++ src/setup.c
@@ -6,7 +6,7 @@
 /*   By: psebasti <sebpalluel@free.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/20 16:46:11 by psebasti          #+#    #+#             */
-/*   Updated: 2017/04/11 17:54:54 by psebasti         ###   ########.fr       */
+/*   Updated: 2017/10/03 11:47:04 by psebasti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,7 +63,7 @@ static int		ft_setup_map_and_mlx(t_setup *setup)
 			return (0);
 		MAP = map;
 	}
-	MLX = ft_initwindow("fdf", (size_t)setup->width, (size_t)setup->height);
+	MLX = ft_initwindow("fdf", setup->width, setup->height);
 	if (MLX && MAP)
 		return (1);
 	return (0);
